Cleaned up Fortran code (everything is now double precision) and interface.

+++ src/RNG_wrapper.c
@@ -6,8 +6,8 @@ void F77_SUB(getrngstate)(void) { GetRNGstate(); }
 void F77_SUB(putrngstate)(void) { PutRNGstate(); }
 
 /* Note: R's unif_rand returns 0<=x<=1 while FORTRAN's RAND returns 0<=x<1 */
-void F77_SUB(unifrand)(double* x) { 
+void F77_SUB(unifrand)(double* x) {
   do{
-    *x = (double) unif_rand();
+    *x = unif_rand();
   }while(*x >= 1.0 || *x <0.0);
 }
